requirements.txt
requests
Pillow
openai
python-dotenv
fastapi
uvicorn
gradio
